Detect clang-cl as MSVC
The compiler identification checks in src/optick.h currently assume the
user is on a Unix platform is compiling with Clang. This isn't the case
when using Clang's MSVC compatible frontend (clang-cl).

This is a simple fix that checks for MSVC first, which both cl.exe
and clang-cl.exe will identify as.