Fix compilation with USE_OPTICK (0)
Last time I've made a assumption that `OPTICK_UNUSED` was already defined so I didn't bother to test, but t turns out that it was not. So when compiling with USE_OPTICK (0) compilation errors are thrown.
Now I've created a new definition `OPTICK_CAPI_UNUSED` to not collide with `OPTICK_UNUSED` and am using that instead.
This time I tested compilation so hopefully it's ok.